agent/scratch/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "influxdb-client>=1.49.0",
    "kubernetes>=33.1.0",
    "numpy>=2.1.3",
    "prometheus-api-client>=0.6.0",
    "python-dotenv>=1.1.1",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
]